breaking out precondition stuff a little bit.
