+++ buckets-mdapi/src/object.rs
@@ -1,9 +1,7 @@
-// Copyright 2019 Joyent, Inc.
+// Copyright 2020 Joyent, Inc.
 
 use std::error::Error;
 use std::vec::Vec;
-use std::fmt;
-use std::io;
 
 use base64;
 use postgres::types::{FromSql, IsNull, ToSql, Type};
@@ -13,6 +11,7 @@ use tokio_postgres::{accepts, to_sql_checked};
 use uuid::Uuid;
 
 use crate::error::{BucketsMdapiError, BucketsMdapiErrorType};
+use crate::sql;
 use crate::types::{HasRequestId, Hstore, RowSlice, Timestamptz};
 
 pub mod create;
@@ -20,12 +19,6 @@ pub mod delete;
 pub mod get;
 pub mod list;
 pub mod update;
-
-use tokio_postgres::Error as PGError;
-use postgres::Transaction;
-use slog::{crit, Logger};
-use crate::sql;
-use crate::metrics;
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct GetObjectPayload {
@@ -36,7 +29,6 @@ pub struct GetObjectPayload {
     pub request_id: Uuid,
     pub headers: Hstore,
 }
-// is_conditional()?
 
 impl HasRequestId for GetObjectPayload {
     fn request_id(&self) -> Uuid {
@@ -158,151 +150,6 @@ pub(self) fn object_not_found() -> Value {
         BucketsMdapiErrorType::ObjectNotFound,
     ))
     .expect("failed to encode a ObjectNotFound error")
-}
-
-
-pub(self) fn precondition_error(msg: String) -> Value {
-    serde_json::to_value(BucketsMdapiError::with_message(
-        BucketsMdapiErrorType::PreconditionFailedError,
-        msg,
-    ))
-    .expect("failed to encode a PreconditionFailedError error")
-}
-
-// XXX
-//
-// This is basically from:
-//
-//     https://blog.burntsushi.net/rust-error-handling/#defining-your-own-error-type
-//
-// I can't say I understand it all yet.
-//
-// I think `conditional` (or whatever it's going to be called) is going to need to have a custom
-// error type because it could either be a PGError from attempting the initial query, or whatever
-// type the actual conditional comparison failure is going to be.
-//
-// Still not sure on what error to return to the client.  I think as it is this will look like a
-// PostgresError, where I think something like PreconditionFailedError is what I want.  This
-// error could then define some ways of printing the precondition failure into a string suitable
-// for the fast message.
-#[derive(Debug)]
-enum ConditionalError {
-    Pg(PGError),
-    Conditional(BucketsMdapiError),
-}
-
-impl From<PGError> for ConditionalError {
-    fn from(err: PGError) -> ConditionalError {
-        ConditionalError::Pg(err)
-    }
-}
-
-impl From<BucketsMdapiError> for ConditionalError {
-    fn from(err: BucketsMdapiError) -> ConditionalError {
-        ConditionalError::Conditional(err)
-    }
-}
-
-impl fmt::Display for ConditionalError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            ConditionalError::Pg(ref err) => write!(f, "{}", err),
-            ConditionalError::Conditional(ref err) => write!(f, "{}", err),
-        }
-    }
-}
-
-impl Error for ConditionalError {
-    fn description(&self) -> &str {
-        match *self {
-            ConditionalError::Pg(ref err) => err.description(),
-            ConditionalError::Conditional(ref err) => err.description(),
-        }
-    }
-
-    fn cause(&self) -> Option<&Error> {
-        match *self {
-            ConditionalError::Pg(ref err) => Some(err),
-            ConditionalError::Conditional(ref err) => Some(err),
-        }
-    }
-}
-
-// XXX
-//
-// This perhaps could return an Option, but I think it's going to have to be Result at least
-// because of the get request case.  In that case the actual call is exactly the same as what is
-// done in this conditional call, so we might as well just return the object(s) and have the caller
-// skip the separate call to the database.
-pub(self) fn conditional(
-    mut txn: &mut Transaction,
-    items: &[&dyn ToSql],
-    vnode: u64,
-    headers: &Hstore,
-    metrics: &metrics::RegisteredMetrics,
-    log: &Logger,
-) -> Result<(), ConditionalError> {
-    // XXX
-    //
-    // This should be some kind of `.isConditional()` call on some part of the request.  Probably
-    // the headers.
-    if !headers.contains_key("if-match") {
-        crit!(log, "if-match not found; returning");
-        return Ok(());
-    }
-
-    let rows = sql::txn_query(
-        sql::Method::ObjectGet,
-        &mut txn,
-        sql::get_sql(vnode).as_str(),
-        items,
-        metrics,
-        log,
-    )?;
-
-    // XXX
-    //
-    // What happens on 0 rows, like a precondition on a non-existent object?
-    //
-    // What about >1 rows?  Do we need to handle all cases like `response` does?
-
-    crit!(log, "got {} rows from conditional query", rows.len());
-
-    if rows.len() == 1 {
-        let row = &rows[0];
-
-        let etag: Uuid = row.get("id");
-
-        // XXX
-        //
-        // Need to figure out properly getting into Some(Some("x"))
-        let x = headers.get("if-match").unwrap().clone();
-        let y = x.unwrap();
-
-        // XXX
-        //
-        // This could be a comma separated list of values.  Or they could be weak comparisons.
-        // Or...
-        //
-        // Should this headers object be a new type with some fancy methods for getting at this
-        // information?
-        let if_match_id = Uuid::parse_str(y.as_str()).unwrap();
-
-        if etag == if_match_id {
-            crit!(log, "precondition success: want:{} / got:{}", if_match_id, etag);
-        } else {
-            let msg = format!("if-match {} didn't match etag {}", if_match_id, etag);
-            crit!(log, "{}", msg);
-            //return Err(ConditionalError::Conditional(io::Error::new(io::ErrorKind::Other, msg)));
-            return Err(ConditionalError::Conditional(BucketsMdapiError::with_message(
-                BucketsMdapiErrorType::PreconditionFailedError,
-                msg,
-            )));
-        }
-    }
-
-    Ok(())
-
 }
 
 pub(self) fn response(
