+++ buckets-mdapi/tests/rpc_handlers.rs
@@ -6,7 +6,7 @@ use std::path::Path;
 use std::process::Command;
 use std::sync::Mutex;
 
-use slog::{crit, error, info, o, Drain, Level, LevelFilter, Logger};
+use slog::{error, info, o, Drain, Level, LevelFilter, Logger};
 use url::Url;
 use uuid::Uuid;
 
@@ -493,8 +493,6 @@ fn verify_rpc_handlers() {
     assert_eq!(get_object_unwrapped_result.name, object);
     assert_eq!(&get_object_unwrapped_result.content_type, "text/html");
 
-    crit!(log, "--- end --- good if-match test");
-
     // Try get object with "if-match: wrongETag"
     let request_id = Uuid::new_v4();
     let mut headers = HashMap::new();
@@ -534,8 +532,6 @@ fn verify_rpc_handlers() {
         ),
     );
 
-    crit!(log, "--- end --- bad if-match test");
-
     // Get object with "if-match: *"
 
     // Delete object
@@ -566,8 +562,6 @@ fn verify_rpc_handlers() {
     assert_eq!(&delete_object_response[0].bucket_id, &bucket_id);
     assert_eq!(&delete_object_response[0].name, &object);
 
-    crit!(log, "--- end --- delete object test");
-
     // Read object again and verify it is not found
     get_object_result =
         util::handle_msg(&get_object_fast_msg, &pool, &metrics, &log);
@@ -584,8 +578,6 @@ fn verify_rpc_handlers() {
         BucketsMdapiError::new(BucketsMdapiErrorType::ObjectNotFound)
     );
 
-    crit!(log, "--- end --- read after delete");
-
     // Delete the object again and verify it is not found
     delete_object_result =
         util::handle_msg(&delete_object_fast_msg, &pool, &metrics, &log);
@@ -601,8 +593,6 @@ fn verify_rpc_handlers() {
         delete_object_response_result.unwrap(),
         BucketsMdapiError::new(BucketsMdapiErrorType::ObjectNotFound)
     );
-
-    crit!(log, "--- end --- delete after delete");
 
     // List buckets and confirm none are found
